[feature] Logout without breaking auto login #171
If users chooses to auto-login at the next visit, the logout action
is performed without removing the session cookie.

Closes #171

+++ client/actions/logout.js
@@ -1,10 +1,12 @@
 import {SET_AUTHENTICATION_STATUS} from "../constants/action-types";
 
-const logout = (cookies, orgSlug) => {
-  cookies.remove(`${orgSlug}_auth_token`, {path: "/"});
-  cookies.remove(`${orgSlug}_username`, {path: "/"});
-  cookies.remove(`${orgSlug}_macaddr`, {path: "/"});
-  sessionStorage.clear();
+const logout = (cookies, orgSlug, userAutoLogin = false) => {
+  if (!userAutoLogin) {
+    cookies.remove(`${orgSlug}_auth_token`, {path: "/"});
+    cookies.remove(`${orgSlug}_username`, {path: "/"});
+    cookies.remove(`${orgSlug}_macaddr`, {path: "/"});
+    sessionStorage.clear();
+  }
 
   return {
     type: SET_AUTHENTICATION_STATUS,
